move constants from utils to constants folder
