+++ src/views/SignupPage/SignupPage.jsx
@@ -12,7 +12,7 @@ import Check from '@material-ui/icons/Check'
 import signupPageStyle from '../../jss/material-kit-pro-react/views/signupPageStyle.jsx'
 import styles from '../../jss/material-kit-pro-react/views/componentsSections/preFooter.jsx'
 // constant
-import { USERNAME, EMAIL, PASSWORD, NEUTRAL } from '../../utils/constants'
+import { USERNAME, EMAIL, PASSWORD, NEUTRAL } from '../../constants/general'
 // image
 import image from '../../../static/img/bg7.jpg'
 // core components
